# potential perf improvements in CustomWebUtility