--- a/tests/PerfTest/ViewEngineBuild.fs
+++ b/tests/PerfTest/ViewEngineBuild.fs
@@ -1,10 +1,6 @@
 ﻿namespace PerfTest
 
-open System.Net
-open System.Text
 open BenchmarkDotNet.Attributes
-open Oxpecker.ViewEngine
-open Oxpecker.ViewEngine.Tools
 
 module OxpeckerViewBuild =
     open Oxpecker.ViewEngine
@@ -70,16 +66,11 @@ type ViewEngineBuild() =
     // | BuildFalcoView    | 836.7 ns | 6.20 ns | 5.49 ns | 0.7763 |   3.17 KB |
 
 
-    let testText = (Seq.init 10000 (fun _ -> "Some long error text")) |> String.concat "<script/>"
+    [<Benchmark>]
+    member this.BuildOxpeckerView() = OxpeckerViewBuild.get()
 
     [<Benchmark>]
-    member this.WebUtility() =
-        let sb = StringBuilder()
-        WebUtility.HtmlEncode testText
-        |> sb.Append
-        |> ignore
+    member this.BuildGiraffeView() = GiraffeViewBuild.get()
 
     [<Benchmark>]
-    member this.Tools() =
-        let sb = StringBuilder()
-        CustomWebUtility.htmlEncode testText sb
+    member this.BuildFalcoView() = FalcoViewBuild.get()