Made some changes that potentially improves perf on my machine at least.
Second is new version.

Skip test is test where no chars need to be encoded.
```
BenchmarkDotNet v0.14.0, Windows 10 (10.0.19045.5131/22H2/2022Update)
Intel Core i9-9900K CPU 3.60GHz (Coffee Lake), 1 CPU, 8 logical and 8 physical cores
.NET SDK 9.0.100
  [Host]     : .NET 9.0.0 (9.0.24.52809), X64 RyuJIT AVX2 DEBUG
  DefaultJob : .NET 9.0.0 (9.0.24.52809), X64 RyuJIT AVX2


| Method          | Mean     | Error   | StdDev  | Allocated |
|---------------- |---------:|--------:|--------:|----------:|
| htmlEncode      | 676.0 us | 9.08 us | 4.50 us |      24 B |
| htmlEncode2     | 593.9 us | 5.34 us | 4.73 us |      24 B |
| htmlEncodeSkip  | 188.5 us | 1.23 us | 0.96 us |      24 B |
| htmlEncode2Skip | 132.8 us | 1.91 us | 1.79 us |      24 B |

```